Added Raindrops and Branch
Added Raindrops to the scemce to have them fall and bounce off of the mushrooms. The raindrops re going to fall off the newly added branch that I am going to texture and possibly add some leaves to if I have time. Then the raindrops could fall off the leaf. Depends on how much time I have to finish it.